Moving bootstrap file to test-export so npm actually publishes it.
