+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@department-of-veterans-affairs/appeals-frontend-toolkit",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Build tools and React components for the Appeals frontends",
   "main": "index.js",
   "repository": "git@github.com:department-of-veterans-affairs/appeals-frontend-toolkit.git",
