+++ test/karma/test-index.js
@@ -6,7 +6,7 @@ import chai from 'chai';
 import sinonChai from 'sinon-chai';
 chai.use(sinonChai);
 
-import '../bootstrap';
+import '../../test-export/bootstrap';
 
 // require all modules ending in "-test" from the
 // current directory and all subdirectories
